chore: add missing user logic for delete user method

+++ src/main/java/com/roy/morago/service/user/UserService.java
@@ -43,7 +43,12 @@ public class UserService {
         userRepository.save(user);
     }
 
+    @Transactional
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("User not found: " + id); // update to use custom exception
+        }
     }
 }
